refactor(builder): refactor build functions
