+++ libs/builder/helpers/angular/get-component-assets.ts
@@ -1,7 +1,7 @@
 import {ClassDeclaration} from 'ts-morph';
 
 import {NgDocAsset} from '../../interfaces';
-import {buildFileAsset} from '../build-assets';
+import {buildFileAsset} from '../build-file-assets';
 import {getComponentSourceFiles} from './get-component-source-files';
 
 /**
