+++ libs/builder/helpers/build-file-assets.ts
